Add controls to frmAddItems and Created ManageLocations
frmManageLocations is used for when an item will have multiple locations. A button will appear when that option is selected. Intention is to take you to that child form, and take you back to AddItems when you save/cancel location changes.

To do:
Set Manage Locations button to take you to frmMultipleLocaitons.
Set Save/Cancel to return you to frmAddItems.
frmMultipleLocations will display the current item's name in the title label.